Migrate CoreReplicationAPI annotations to 'api' module
