+++ solr/api/src/java/org/apache/solr/client/api/model/IndexVersionResponse.java
@@ -14,23 +14,27 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.apache.solr.client.api.model;
 
-package org.apache.solr.client.api.util;
+import com.fasterxml.jackson.annotation.JsonProperty;
 
-public class Constants {
-  private Constants() {
-    /* Private ctor prevents instantiation */
+/** Response body for the `GET /api/cores/coreName/replication/indexversion` API */
+public class IndexVersionResponse extends SolrJerseyResponse {
+
+  @JsonProperty("indexversion")
+  public Long indexVersion;
+
+  @JsonProperty("generation")
+  public Long generation;
+
+  @JsonProperty("status")
+  public String status;
+
+  public IndexVersionResponse() {}
+
+  public IndexVersionResponse(Long indexVersion, Long generation, String status) {
+    this.indexVersion = indexVersion;
+    this.generation = generation;
+    this.status = status;
   }
-
-  public static final String INDEX_TYPE_PATH_PARAMETER = "indexType";
-  public static final String INDEX_NAME_PATH_PARAMETER = "indexName";
-  public static final String INDEX_PATH_PREFIX =
-      "/{" + INDEX_TYPE_PATH_PARAMETER + ":cores|collections}/{" + INDEX_NAME_PATH_PARAMETER + "}";
-
-  public static final String CORE_NAME_PATH_PARAMETER = "coreName";
-
-  public static final String OMIT_FROM_CODEGEN_PROPERTY = "omitFromCodegen";
-  public static final String GENERIC_ENTITY_PROPERTY = "genericEntity";
-
-  public static final String BINARY_CONTENT_TYPE_V2 = "application/vnd.apache.solr.javabin";
 }
